Add Code.stories.tsx + move Code component
