+++ src/pages/[locale]/examples/built-in-features/static-i18n.tsx
@@ -21,7 +21,7 @@ import BuiltInFeaturesSidebar from '../../../../components/doc/BuiltInFeaturesSi
 import DocPage from '../../../../components/doc/DocPage';
 import DefaultLayout from '../../../../components/pageLayouts/DefaultLayout';
 import DisplayOnBrowserMount from '../../../../components/rehydration/DisplayOnBrowserMount';
-import Code from '../../../../components/utils/Code';
+import Code from '../../../../components/dataDisplay/Code';
 import ExternalLink from '../../../../components/utils/ExternalLink';
 import useI18n, { I18n } from '../../../../hooks/useI18n';
 import { CommonServerSideParams } from '../../../../types/nextjs/CommonServerSideParams';
